refactor: extract ObjectStateRegistry to separate file
- Split object_state.py (2507 lines) into:
  - object_state_registry.py (1182 lines): ObjectStateRegistry singleton
  - object_state.py (1350 lines): FieldProxy + ObjectState
- Updated all 22 import statements across pyqt_gui
- Fixed indentation issues from sed replacement in 4 files
- Verified OpenHCS runs successfully
